ENH: Add separate_axes option, apidoc fix, docs, merge tests
- Added separate_axes parameter to plot_expect for subplot-per-e_op
- Fixed apidoc so plot_expect and plot_photocurrent appear in
  McResult and NmmcResult documentation
- Added Users Guide section on visualising solver results
- Moved tests from test_result_plot.py into test_results.py
